mg still does not work, but solved (?hopefully) issues on assembly for more than one surfaec integrak

+++ mars_base.hpp
@@ -8,7 +8,7 @@
  * M.A.R.S Mesh Adaptive Refinement for Simplical meshes
  */
 namespace mars {
-    using Real    = double;
+    using Real    = long double;
     using Integer = long;
     static const long INVALID_INDEX = -1;
 }
